+++ app/controllers/application_controller.rb
@@ -11,4 +11,9 @@ class ApplicationController < ActionController::Base
       devise_parameter_sanitizer.permit :sign_in, keys: added_attrs
   end
 
+  def current_user
+    ActiveDecorator::Decorator.instance.decorate(super) if super.present?
+    super
+  end
+
 end
