+++ app/decorators/tweet_decorator.rb
@@ -1,0 +1,9 @@
+module TweetDecorator
+  def display_created_at
+    I18n.l(self.created_at, format: :default)
+  end
+
+  def like_count
+    likes.count
+  end
+end
