+++ app/decorators/user_decorator.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module UserDecorator
+  
+  def avatar_image
+    if profile&.avatar&.attached?
+      profile.avatar
+    else
+      'default-avatar.png'
+    end
+  end
+
+
+end
