+++ app/models/tweet.rb
@@ -22,11 +22,4 @@ class Tweet < ApplicationRecord
   has_one_attached :eyecatch
   has_rich_text :content
 
-  def display_created_at
-    I18n.l(self.created_at, format: :default)
-  end
-
-  def like_count
-    likes.count
-  end
 end
