+++ app/models/user.rb
@@ -62,14 +62,7 @@ class User < ApplicationRecord
   def prepare_profile
     profile || build_profile
   end
-
-  def avatar_image
-    if profile&.avatar&.attached?
-      profile.avatar
-    else
-      'default-avatar.png'
-    end
-  end
+  
 
   private
   def get_user_id(user)
